Updates by INCOSE
Update to latest work by INCOSE working group. Here are the notes:

The CSRM residing in GitHub is the “official” CSRM.
It was downloaded and “refined” as described below.
It will be “returned” to GitHub
The CSRM Profile was not touched.
2.  CSRM Download and Open.    2024-09-07
The purpose of this memo is to establish the date of the CSRM in GitHub then download the CSRM for updates as described in CSRM Package Deletions
3.  CSRM HTML Navigation.      2024-09-25
This provides the capability for a user to explore the CSRM with a browser instead of Cameo.
4.  CSRM Package Deletions.      2024-09-13
The purpose on this memo is illustrate and agree on csrm package deletions.
Then carryout spot check to verify the CSRM was updated
5.  Change Behavior Link to Package Diagram.    2024-09-16
The purpose of this memo is to verify the CSRM was updated accordingly
6.  Validation and Verification Activities - Population Instructions Fixup.    2024-10-12
The purpose of this memo is to verify that the Validation Activities package diagram and Verification Activities package diagram have been updated as shown on pages 6 and 7
7.  Validation and Verification Activities.     2024-10-12
The purpose of this memo is to address the population of validation and verification activities tables with respect to technical measure specifications and requirements
The CSRM architecture is sufficient as is.
8.  Technical Measures Methods - Or Not.    2024-10-15
Refer to page 4 for the withdrawal of the Methods document concept and updates to the CSRM
9.  Add New Validation and Verification Activities Buttons.    2024-12-09
The purpose of this memo is to see if the Add New Validation Activity and Add New Verification Activity buttons work
Validation worked.  Verification did not but was fixed as shown below.
 
10.  Containment Tree - Cameo.    2024-10-18
  See Figure 3c.  Delete as shown
11.  Containment Tree - HTLM.    2024-10-18
  See page 3 for delete of Enterprise Example
12.  Update CSRM Quick Look Landing Page.    2024-10-24
13.  Telemetry Processing Subsystem.    2024-11-30
  Add a Telemetry Processing Subsystem to the CSRM.
14.  Telemetry Processing Components - Population.    2024-11-30
  Add a Telemetry Processing Components - Population package to the CSRM
15.  Telemetry Processing Requirements - Table.   2024-11-30.
  Add a Telemetry Processing Requirements Table to the CSRM
16.  Component Placeholders.    2024-11-30
  The placeholders are replaced when creating the mission logical and physical models
17.  Subsystem bdds Uniformity.    2024-11-30
  Purpose of this memo is to provide uniformity in the subsystem bdds
18.  Component bdds Uniformity.    2024-11-30
  Purpose of this memo is to provide uniformity in the component bdds
19.  CSRM: Subsystems, Subsystem Component Packages, and Population Package Diagrams
  Illustrates the relationships of subsystems, components, and populations

NOTE: The XMI version have not been updated, just the Dassault files.